Move help related urls to own apiUrls; a few new shared/config/urls files too.
