+++ src/contact/config/apiUrls.ts
@@ -1,0 +1,6 @@
+const apiUrls = {
+  contact: {
+    token: joinUrl(apiPrefix, 'contact', 'token'),
+    send: joinUrl(apiPrefix, 'contact', 'send'),
+  },
+};
